refactor(fbuild-daemon): split handlers/operations.rs into submodules to satisfy LOC gate
